CID-1401 - Reconcile DA tenant roles when tenant modified (#1921)
* CID-1401 - Reconcile DA tenant roles when tenant modified

* CID-1401 - addressing review comments

* CID-1401 - Updated cleanup logic when tenant is moving between domains per review comments

* CID-1401 - Addressing nits

+++ src/main/java/com/rackspace/idm/domain/service/impl/DefaultTenantService.java
@@ -852,7 +852,7 @@ public class DefaultTenantService implements TenantService {
             throw new NotFoundException(errMsg);
         }
 
-        deleteTenantFromTenantRole(role, tenant);
+        deleteTenantFromTenantRole(role, tenant.getTenantId());
 
         if (endUser instanceof User) {
             //this only applies for users, not federatedusers for now...
@@ -865,7 +865,7 @@ public class DefaultTenantService implements TenantService {
                     try {
                         TenantRole subUserTenantRole = tenantRoleDao.getTenantRoleForUser(subUser, role.getRoleRsId());
                         if (subUserTenantRole != null) {
-                            deleteTenantFromTenantRole(subUserTenantRole, tenant);
+                            deleteTenantFromTenantRole(subUserTenantRole, tenant.getTenantId());
                             atomHopperClient.asyncPost(subUser, AtomHopperConstants.ROLE);
                         }
                     } catch (NotFoundException ex) {
@@ -878,7 +878,7 @@ public class DefaultTenantService implements TenantService {
                 for(FederatedUser subUser : federatedUserDao.getUsersByDomainId(user.getDomainId())) {
                     try {
                         TenantRole subUserTenantRole = tenantRoleDao.getTenantRoleForUser(subUser, role.getRoleRsId());
-                        deleteTenantFromTenantRole(subUserTenantRole, tenant);
+                        deleteTenantFromTenantRole(subUserTenantRole, tenant.getTenantId());
                     } catch (NotFoundException ex) {
                         String msg = String.format("Federated user %s does not have tenantRole %s", user.getId(), role.getName());
                         logger.warn(msg);
@@ -890,12 +890,16 @@ public class DefaultTenantService implements TenantService {
 
     }
 
-    private void deleteTenantFromTenantRole(TenantRole role, Tenant tenant) {
-        role.getTenantIds().remove(tenant.getTenantId());
-        if(role.getTenantIds().size() == 0) {
-            tenantRoleDao.deleteTenantRole(role);
-        } else {
-            tenantRoleDao.updateTenantRole(role);
+    @Override
+    public void deleteTenantFromTenantRole(TenantRole role, String tenantId) {
+        boolean roleContainedTenant = role.getTenantIds().remove(tenantId);
+
+        if (roleContainedTenant) {
+            if(role.getTenantIds().size() == 0) {
+                tenantRoleDao.deleteTenantRole(role);
+            } else {
+                tenantRoleDao.updateTenantRole(role);
+            }
         }
     }
